Allow subrepo URL overrides through mapping file.
Subrepo urls are replaced by the path specified in the subrepo mapping file by default. This generally works because Git allows resolution of relative subrepo urls with respect to the parent repo url. In some cases, this may not yield the most desirable result; for example, if the web UI of the git host does not properly support relative urls for subrepos.

This changeset adds a --subrepo-urlmap switch to hg-fast-export.py that allows mapping of subrepo URLs to something other than the default value. This mapping file is used to map original urls (from hgsub) to new urls (in gitmodules).